refactor: 이미지 업로드에 필요한 상태들을 전역 스토어에 저장 #83
기존에는 Make페이지의 전용 store에 이미지 업로드에 필요한 상태들이 있었고, 하위 컴포넌트들은 pages/make/store에서 그것을 제공해주는 context를 import해서 사용하는 방식이었습니다. 하지만 하위 컴포넌트들은 Edit 페이지에서도 사용되어야 하기에 make 페이지의 스토어에 의존성이 생기면 안됩니다. 따라서 업로드에 필요한 상태들을 전역 상태로 올려, GlobalStore에 저장합니다.
